Add files via upload
This system is a lightweight hospital managment tool built enterly using Object Oriented Programming in Python. It registers patients, manages doctors, and allows appointment bookings with billing. The system keeps data in memory and works through a sinple command-line interface. The purpose of this program is to streamline operations, improve data accuracy, and reduce administratie workload by automating processes such as patient registartion, appointment scheduling, and billing, which allows for better patient care and reductionn of wait times. Some limitations to this program would be the lack of user authentication which might allow for anyone to be able to log into the system. Another would be the lack of reason for a patient cancelling an appointment. In order to run this program, one would need to have Python 3.x installed on their system, save the code and open terminal and then run the program. The program displays a numbered menu, where the user is prompted to enter numbers to select actions like registering patients, adding doctors, booking or cancelling appointments, and viewing schedules. Input fields appear as prompted (patient name, doctor ID). Each action runs immediately after input, with feedback messages printed in the terminal. When the program ends, all data is lost and to restart you'd need to rerun the Python file. Some modifications could be adding the option to save, add user authentication(username/password), or implementing appointment conflict.